requirements.txt
pandas==2.1.3
python-dateutil==2.8.2
spacy==3.1.4
geopy==2.2.0
geopandas==0.10.2
Shapely==1.7.1
sentinelhub==3.4.0
Pillow==8.3.2
matplotlib==3.4.3
numpy==1.21.2
six==1.16.0   # Only if it's necessary
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
requests==2.31.0
pytz==2023.3.post1
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.104.1
h11==0.14.0
httptools==0.6.1
idna==3.4
PyYAML==6.0.1
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
